fix: 필터가 반응형으로 태블릿 필터 및 static 필터로 적용하지않던 문제
